🔧 Code improvement 🔧 Added a ban on navigating through browser history
