+++ package.json
@@ -22,6 +22,7 @@
     "@angular/platform-browser-dynamic": "^15.2.0",
     "@angular/router": "^15.2.0",
     "@ngneat/until-destroy": "^10.0.0",
+    "angular-disable-browser-back-button": "^2.0.0",
     "angular-svg-icon": "^15.0.0",
     "rxjs": "~7.8.0",
     "tslib": "^2.3.0",
