+++ src/app/components/pages/dictionary/dictionary.component.scss
@@ -1,5 +1,5 @@
-@import "src/app/shared/styles/mixins";
-@import "src/app/shared/styles/colors";
+@import "../../../shared/styles/mixins";
+@import "../../../shared/styles/colors";
 
 .dictionary {
   display: flex;
