+++ src/app/components/pages/dictionary/dictionary.component.spec.ts
