+++ src/app/components/pages/main-manu/main-manu.component.scss
@@ -1,4 +1,4 @@
-@import "src/app/shared/styles/mixins";
+@import "../../../shared/styles/mixins";
 
 .menu {
   display: flex;
