+++ src/app/components/pages/main-manu/main-manu.component.ts
