+++ src/app/components/pages/result/result.component.spec.ts
