+++ src/app/components/pages/test/test.component.scss
@@ -1,4 +1,4 @@
-@import "src/app/shared/styles/mixins";
+@import "../../../shared/styles/mixins";
 
 .test {
   flex-direction: column;
