+++ src/app/components/toolbar/toolbar.component.ts
@@ -1,4 +1,4 @@
-import { Component } from '@angular/core';
+import { Component, EventEmitter, Input, Output } from '@angular/core';
 import { SettingsService } from "../../services/settings.service";
 import { ResetService } from "../../services/reset/reset.service";
 import { DialogService } from "../../services/dialog.service";
@@ -10,6 +10,10 @@ import { Router } from "@angular/router";
   styleUrls: ['./toolbar.component.scss']
 })
 export class ToolbarComponent {
+  @Input() showStatisticsBtn: boolean = false;
+  @Input() resetTestBtn: boolean = false;
+  @Output() back: EventEmitter<void> = new EventEmitter<void>();
+
 
   constructor(
     private settingService: SettingsService,
@@ -27,6 +31,6 @@ export class ToolbarComponent {
   }
 
   public backToMainMenu(): void {
-    this.dialogService.openBackToMainMenu(() => this.router.navigate(['main-menu']).then(() => this.resetService.run()));
+    this.back.emit();
   }
 }
