+++ src/app/interfaces/tags.interface.ts
@@ -4,27 +4,27 @@ export enum ETag {
   A = '<a>',
   Abbr = '<abbr>',
   Address = '<address>',
-  Applet = '<applet>',
-  Area = '<area>',
-  Article = '<article>',
-  Aside = '<aside>',
-  Audio = '<audio>',
-  B = '<b>',
-  Base = '<base>',
-  Basefont = '<basefont>',
-  Bdo = '<bdo>',
-  Big = '<big>',
-  Blockquote = '<blockquote>',
-  Body = '<body>',
-  Br = '<br>',
-  Button = '<button>',
-  Canvas = '<canvas>',
-  Caption = '<caption>',
-  Center = '<center>',
-  Cite = '<cite>',
-  Code = '<code>',
-  Col = '<col>',
-  Colgroup = '<colgroup>',
+  // Applet = '<applet>',
+  // Area = '<area>',
+  // Article = '<article>',
+  // Aside = '<aside>',
+  // Audio = '<audio>',
+  // B = '<b>',
+  // Base = '<base>',
+  // Basefont = '<basefont>',
+  // Bdo = '<bdo>',
+  // Big = '<big>',
+  // Blockquote = '<blockquote>',
+  // Body = '<body>',
+  // Br = '<br>',
+  // Button = '<button>',
+  // Canvas = '<canvas>',
+  // Caption = '<caption>',
+  // Center = '<center>',
+  // Cite = '<cite>',
+  // Code = '<code>',
+  // Col = '<col>',
+  // Colgroup = '<colgroup>',
   // Datalist = '<datalist>',
   // Dd = '<dd>',
   // Del = '<del>',
