+++ src/styles.scss
@@ -27,6 +27,8 @@ input, button {
 
   width: 27px;
   height: 27px;
+  min-width: 27px;
+  min-height: 27px;
   box-sizing: border-box;
   cursor: pointer;
   user-select: none;
